pyproject.toml
[tool.poetry]
name = "python-poetry-docker-example"
version = "0.1.0"
description = ""
authors = ["Someone <someone@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"

[tool.poetry.dev-dependencies]
pandas = "^1.5"
sqlalchemy = "^1.4"
pyarrow = "^11.0.0"
psycopg2-binary = "^2.9"
pyspark = "^3.4.0"

[build-system]
requires = ["poetry-core>=1.2.0"]
build-backend = "poetry.core.masonry.api"